==> scripts/02_create_shps.ipynb <==
<br>

# Introdução

Este _script_ tem objetivo de fazer a transformação das tabelas em mapas.


#!pip3 install geopandas --upgrade
#!pip3 install open_geodata --upgrade
#!pip3 install selenium
!pip3 install pyarrow

import geopandas as gpd
import pandas as pd
from open_geodata import geo
from paths import output_path_geojson, output_path_gpkg, output_path_tab

<br>

## Join Data

Inicialmente lemos os dados tabulares.


# Lê o arquivo csv com o nome dos municípios
df = pd.read_csv(
    output_path_tab / 'tab_cetesb_municipios.csv',
)

# Results
df.info()
df.head(2)

<br>

Criamos um geodatafram com a delimitação dos municípios no estado de São Paulo.


gdf = geo.load_dataset('geo.sp.sp_250k_wgs84')

# # Lê o arquivo csv com o nome dos municípios
gdf = gdf.drop(['municipio_nome'], axis=1, inplace=False)

# Results
gdf.info()
gdf.head()

# gdf['id_municipio'] = gdf['id_municipio'].astype(int)
gdf['geometry'] = gdf.simplify(0.0015)

# Results
gdf.info()
gdf.head()

<br>

E concatenamos as informações.


# Merge
gdf = gdf.merge(df, on='id_municipio', how='left')

# Results
gdf.info()
gdf.head()

<br>

## Save geojson

Salvamos os resultados.


gdf.to_file(
    output_path_gpkg / 'sp_cetesb.gpkg',
    layer='sp_cetesb',
    driver='GPKG',    
)

gdf.to_file(
    output_path_geojson / 'sp_cetesb.geojson',
    driver='GeoJSON',
    encoding='utf-8',
)

# Results
gdf.head()

gdf.explore(column='agencia')
