==> scripts/03_convert.ipynb <==
<br>

# Introdução


#!pip3 install open-geodata --upgrade

from open_geodata import converts
from paths import output_path_gpkg, output_path_geo

<br>

## GPKC para 7ZIP


converts.files.convert_to_7zip(output_path_gpkg, output_path_geo)
